added code for pose reception from camera

+++ swarm_camera/pid.ino
@@ -1,15 +1,15 @@
-const double kpA=1000;
-const double kiA=22;
-const double kdA=1.3;
+const float kpA=1000;
+const float kiA=22;
+const float kdA=1.3;
 
-const double kpB=1000;
-const double kiB=22;
-const double kdB=1.3;
+const float kpB=1000;
+const float kiB=22;
+const float kdB=1.3;
 
 void pwm_pid(){
  errorA = set_rpmA - curr_rpmA;
  IA+=errorA;
-  double x = 0;
+  float x = 0;
  if((kpA*errorA + kiA*IA + kdA*(errorA-preverrorA))<0)
   x = 0;
  else
@@ -17,7 +17,7 @@ void pwm_pid(){
  PWMA=map(x, 0, max_rpmA*kpA, 0, 255);
  preverrorA=errorA;
 
-double y = 0;
+float y = 0;
  errorB = set_rpmB - curr_rpmB;
 IB+=errorB;
 if((kpB*errorB + kiB*IB + kdB*(errorB-preverrorB))<0)
